backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
httpx>=0.25.0
groq>=0.4.0
pillow>=10.2.0
dnspython>=2.4.0
certifi>=2024.2.2
aiosqlite>=0.19.0
